# tests/baseline_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MH Gateway",
    "summary": "\u7f16\u6392\u7f51\u5173 \u2014 \u573a\u666f\u52a0\u8f7d\u3001Agent \u8def\u7531\u3001\u4e8b\u4ef6\u6d41\u5f52\u96c6",
...
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/eval/batches": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "List Batches",
        "operationId": "list_batches_api_v1_eval_batches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Create Batch Eval",
        "operationId": "create_batch_eval_api_v1_eval_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchEvalRequestSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eval/batches/{batch_id}": {
      "get": {
        "tags": [
          "eval"
        ],
        "summary": "Get Batch",
        "operationId": "get_batch_api_v1_eval_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/eval/batches/{batch_id}/cancel": {
      "post": {
        "tags": [
          "eval"
        ],
        "summary": "Cancel Batch",
        "operationId": "cancel_batch_api_v1_eval_batches__batch_id__cancel_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BatchEvalRequestSchema": {
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/EvalQuestionSchema"
            },
            "type": "array",
            "minItems": 1,
            "title": "Questions"
          },
          "llm_provider": {
            "type": "string",
            "title": "Llm Provider"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model"
          },
          "max_concurrency": {
            "type": "integer",
            "title": "Max Concurrency",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "questions",
          "llm_provider",
          "llm_model"
        ],
        "title": "BatchEvalRequestSchema"
      },
...
      "EvalQuestionSchema": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "input_text",
          "scenario_id",
          "agent_name"
        ],
        "title": "EvalQuestionSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}